+++ marketing-analytics/activation/data-tasks-coordinator/src/tasks/internal/status_check_task.js
@@ -19,28 +19,48 @@
 
 'use strict';
 
-const {TaskLog, FIELD_NAMES} = require('../../task_log/task_log_dao.js');
+const { utils: { wait } } = require('@google-cloud/nodejs-common');
+const {
+  TaskLogStatus,
+  TaskLogDao,
+  FIELD_NAMES,
+} = require('../../task_log/task_log_dao.js');
+const { ErrorHandledStatus } = require('../error/error_handled_status.js');
+const { RetryableError } = require('../error/retryable_error.js');
 const {BaseTask} = require('../base_task.js');
 
 /**
- * Status of manual asynchronous tasks.
+ * Status of checked tasks.
  * @enum {string}
  */
-const EXTERNAL_JOB_STATUS = Object.freeze({
+const CHECKED_TASK_STATUS = Object.freeze({
   DONE: 'DONE',
   RUNNING: 'RUNNING',
+  RETRIED: 'RETRIED',
   ERROR: 'ERROR',
   ERROR_IGNORED: 'ERROR_IGNORED',
 });
 
 /**
+ * When the number of async tasks is lower, it is very likely it starts handle
+ * the mutilple-layered embedded tasks.
+ * @const{number}
+ */
+const EXTRA_CHECK_THRESHOLD = 10;
+
+/** @const{number} Maximum times of extra checks. */
+const EXTRA_CHECK_TIMES = 5;
+
+/**
+ * Log data structure for a check of a task.
+ *
  * @typedef {{
  *   taskLogId: (string|number),
- *   status: EXTERNAL_JOB_STATUS,
+ *   status: CHECKED_TASK_STATUS,
  *   messageId: (string|undefined),
  * }}
  */
-let TaskLogStatus;
+let TaskCheckLog;
 
 /**
  * Internal Task to update all running manual asynchronous tasks' status:
@@ -48,6 +68,9 @@ let TaskLogStatus;
  *   1. If completed, sends out a message to notify Sentinel to finish it;
  *   2. If there is an error, logs the error message to TaskLog;
  *   3. If it is not completed, then do nothing to that TaskLog.
+ *
+ * Updated 2022/06/08:
+ *   Add handling process for left over tasks due to Cloud Functions' crashes.
  *
  * To start this task, send a message to the 'task monitor' topic with the
  * parameters includes '{intrinsic: "status_check"}'.
@@ -76,76 +99,162 @@ class StatusCheckTask extends BaseTask {
   }
 
   /**
-   * @return {!Promise<{results:!Array<!TaskLogStatus>}>}
+   * @return {!Promise<{results:!Array<!TaskCheckLog>}>}
    * @override
    */
   async doTask() {
     const startTime = new Date();
     const filter = {property: FIELD_NAMES.REGULAR_CHECK, value: true};
-    const taskLogs = await this.options.taskLogDao.list([filter]);
+    const asyncTaskLogs = await this.options.taskLogDao.list([filter]);
+    const crashedTaskLogs = await this.getCrashedTasks_();
+    const taskLogs = asyncTaskLogs.concat(crashedTaskLogs);
     // If there is no task need to be checked, try to pause the cron job.
     if (taskLogs.length === 0) {
       await this.pauseCronjobIfNoNewTasks_(startTime);
       return {results: []};
     }
-    /**
-     * Function to update status of TaskLogs and returns the results.
-     * @param {!Array<TaskLogStatus>} previous
-     * @param {number|string} taskLogId
-     * @param {TaskLog} taskLog
-     * @return {!Promise<!Array<TaskLogStatus>>}
-     */
-    const reduceFn = async (previous, {id: taskLogId, entity: taskLog}) => {
+    const extraCheck = asyncTaskLogs.length < EXTRA_CHECK_THRESHOLD;
+    // To reduce the impact of concurrent requests, using 'reduce' to check the
+    // tasks one by one. If 'reduce' causes a performance issue, then change to
+    // use 'apiSpeedControl' nodejs-common.
+    const results = await taskLogs.reduce(this.getReduceFn(), []);
+    if (!extraCheck) return { results };
+    return { results: await this.extraCheck_(asyncTaskLogs, results) };
+  }
+  /**
+   * Gets the reduce function to update status of TaskLogs and returns results.
+   * @param {!Array<TaskCheckLog>} previous
+   * @param {number|string} taskLogId
+   * @param {TaskLog} taskLog
+   * @return {!Promise<!Array<TaskCheckLog>>}
+   */
+  getReduceFn() {
+    return async (previous, { id: taskLogId, entity: taskLog }) => {
       const results = await previous;
       const currentResult = await this.updateTaskLog_(taskLogId, taskLog);
       results.push(currentResult);
       return results;
     };
-    // To reduce the impact of concurrent requests, using 'reduce' to check the
-    // tasks one by one. If 'reduce' causes a performance issue, then change to
-    // use 'apiSpeedControl' nodejs-common.
-    return {results: await taskLogs.reduce(reduceFn, [])};
+  }
+
+  /**
+   * Returns an array of crashed tasks.
+   * Currently, only check those tasks in 'FINISHING' status.
+   * @return {!Array<{id:string,entity:!TaskLog}>}
+   * @private
+   */
+  async getCrashedTasks_() {
+    const startTime = new Date();
+    const filter = { property: 'status', value: TaskLogStatus.FINISHING };
+    const taskLogs = await this.options.taskLogDao.list([filter]);
+    const now = new Date();
+    return taskLogs.filter(({ id: taskLogId, entity: taskLog }) => {
+      if (now - taskLog.prefinishTime.toDate() > 9 * 60 * 1000) {
+        this.logger.warn('Got a timeout finishing task:', taskLog);
+        return true;
+      }
+      this.logger.info('In time finishing task:', taskLogId);
+      return false;
+    });
   }
 
   /**
    * Checks and returns one taskLog status.
    * @param {string} taskLogId
    * @param {!TaskLog} taskLog
-   * @return {!TaskLogStatus}
+   * @return {!TaskCheckLog}
    * @private
    */
   async updateTaskLog_(taskLogId, taskLog) {
     const taskConfigId = taskLog.taskId;
     const parameters = JSON.parse(taskLog.parameters);
     try {
-      const task = await this.prepareExternalTask(taskConfigId, parameters);
-      const done = await task.isDone();
-      if (done) return this.sendFinishTaskMessage_(taskLogId);
-      return {taskLogId, status: EXTERNAL_JOB_STATUS.RUNNING,};
+      if (taskLog.status === TaskLogStatus.FINISHING) {
+        throw new RetryableError('Cloud Functions crashed.');
+      } else {
+        const task = await this.prepareExternalTask(taskConfigId, parameters);
+        const done = await task.isDone();
+        if (done) return this.sendFinishTaskMessage_(taskLogId);
+        return { taskLogId, status: CHECKED_TASK_STATUS.RUNNING, };
+      }
     } catch (error) {
       this.logger.error(`Error in task: ${taskLogId}`, error);
-      const taskConfig = await this.options.taskConfigDao.load(taskConfigId);
-      const errorOptions = taskConfig.errorOptions || {};
-      if (errorOptions.ignoreError !== true) {
-        await this.options.taskLogDao.saveErrorMessage(taskLogId, error);
-        return {taskLogId, status: EXTERNAL_JOB_STATUS.ERROR,};
+      const [errorHandledStatus] =
+        await this.taskManager.handleFailedTaks(taskLogId, taskLog, error);
+      switch (errorHandledStatus) {
+        case ErrorHandledStatus.RETRIED:
+          return { taskLogId, status: CHECKED_TASK_STATUS.RETRIED, };
+        case ErrorHandledStatus.FAILED:
+          return { taskLogId, status: CHECKED_TASK_STATUS.ERROR, };
+        case ErrorHandledStatus.IGNORED:
+          this.logger.debug('Ignore error task: ', taskLogId);
+          const taskConfig = await this.options.taskConfigDao.load(taskConfigId);
+          if (taskConfig.next) {
+            await this.taskManager.startTasks(taskConfig.next,
+              { [FIELD_NAMES.PARENT_ID]: taskLogId },
+              taskLog.parameters); // Error task pass out origin parameters.
+          }
+          return { taskLogId, status: CHECKED_TASK_STATUS.ERROR_IGNORED, };
       }
-      this.logger.debug('Ignore error task: ', taskLogId);
-      await this.options.taskLogDao.saveErrorMessage(taskLogId, error, true);
-      if (taskConfig.next) {
-        await this.taskManager.startTasks(taskConfig.next,
-            {[FIELD_NAMES.PARENT_ID]: taskLogId},
-            taskLog.parameters); // Error task pass out origin parameters.
-      }
-      return {taskLogId, status: EXTERNAL_JOB_STATUS.ERROR_IGNORED,};
     }
   }
 
   /**
+   * Run extra checks those running task with the conditions:
+   *  1) some tasks are done(including failed);
+   *  2) some tasks are still running.
+   * It waits 10 seconds between each round extra checks.
+   * It runs up to 5 times and merges the results to the original one.
+   * @param {Array<{id:string,entity:!TaskLog}>} taskLogs The array of TaskLog
+   *     that need to be extra checked.
+   * @param {!Array<TaskCheckLog>} results The results of the original check.
+   * @return {!Array<TaskCheckLog>} The combination of the original check and
+   *     the extra checks.
+   * @private
+   */
+  async extraCheck_(taskLogs, results) {
+    const mappedTaskLogs = {};
+    taskLogs.forEach(({ id: taskLogId, entity: taskLog }) => {
+      mappedTaskLogs[taskLogId] = taskLog;
+    });
+    let checkedResults = results.filter(
+      ({ taskLogId }) => !!mappedTaskLogs[taskLogId]
+    );
+    let checkedTimes = 0;
+    do {
+      const runningTaskLogIds = checkedResults
+        .filter(({ status }) => status === CHECKED_TASK_STATUS.RUNNING)
+        .map(({ taskLogId }) => taskLogId);
+      this.logger.info(`Extra check ${checkedTimes}`, runningTaskLogIds);
+      if (runningTaskLogIds.length === 0) break;
+      await wait(10000);
+      checkedResults = await runningTaskLogIds.map(
+        (id) => ({ id, entity: mappedTaskLogs[id] })
+      ).reduce(this.getReduceFn(), []);
+      results.push(...checkedResults);
+    } while (this.hasDoneTasks_(checkedResults) && ++checkedTimes < EXTRA_CHECK_TIMES)
+    return results;
+  }
+
+  /**
+   * Returns whether there is a task done.
+   * @param {!Array<TaskCheckLog>} results
+   * @returns
+   * @private
+   */
+  hasDoneTasks_(results) {
+    return results.some(({ status }) => {
+      return status === CHECKED_TASK_STATUS.DONE ||
+        status === CHECKED_TASK_STATUS.ERROR ||
+        status === CHECKED_TASK_STATUS.ERROR_IGNORED;
+    });
+  }
+
+  /**
    * Sends a 'finish task' message to Pub/Sub to notify Sentinel and returns the
-   * TaskLogStatus.
+   * TaskCheckLog.
    * @param {string} taskLogId
-   * @return {!TaskLogStatus}
+   * @return {!TaskCheckLog}
    * @private
    */
   async sendFinishTaskMessage_(taskLogId) {
@@ -156,7 +265,7 @@ class StatusCheckTask extends BaseTask {
     return {
       taskLogId,
       messageId,
-      status: EXTERNAL_JOB_STATUS.DONE,
+      status: CHECKED_TASK_STATUS.DONE,
     };
   }
 
@@ -180,6 +289,6 @@ class StatusCheckTask extends BaseTask {
 }
 
 module.exports = {
-  EXTERNAL_JOB_STATUS,
+  CHECKED_TASK_STATUS,
   StatusCheckTask,
 };